make chain_id u32

+++ core/src/consensus/consensus_trait.rs
@@ -43,7 +43,7 @@ pub trait ConsensusGraphTrait: Send + Sync {
 
     fn latest_confirmed_epoch_number(&self) -> u64;
 
-    fn best_chain_id(&self) -> u64;
+    fn best_chain_id(&self) -> u32;
 
     fn best_block_hash(&self) -> H256;
